requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap4==2024.10
Django==5.1.4
django-crispy-forms==2.3
django-taggit==6.1.0
pillow==11.1.0
reportlab==4.2.5
sqlparse==0.5.3
Unidecode==1.3.8
